# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Word transcription",
        "description": "Api para fazer a transcrição da palavra em inglês",
        "version":"1.0"
    },
    "servers":[
        {
            "url":"http://localhost:3000"
        }
    ],
    "paths":{
...
        "/verbete/{palavra}/xml":{
            "get":{
                "description":"É passado uma palavra e retorna o verbete dela em português",
                "tags":["Word"],
                "parameters":{
                    
                    "palavra":"palavra",
                    "in":"query"
                    
                },
                "responses":{
                    "200":"Ok",
                    "400":"Vebete não encontrado"
                }
            }
        }
    },
...
}